feat: rename functions and files to be consistent with each other (#10358)
